tensorflow_implementation/requirements.txt
h5py==2.10.0
Keras==2.1.6
matplotlib==3.0.3
nltk==3.2.5
numpy==1.17.3
opencv-python==3.4.0.12
pandas==0.21.0
Pillow==6.2.0
protobuf==3.5.0.post1
scikit-learn==0.19.1
scipy==1.3.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3


